+++ src/Application/Command/Happening/Webinar/Poll/PollHappeningMismatchException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Command\Happening\Webinar\Poll;
+
+class PollHappeningMismatchException extends \LogicException
+{
+}