+++ src/Application/Query/Happening/Admin/HappeningViewQueryHandler.php
@@ -56,7 +56,8 @@ class HappeningViewQueryHandler
             $happening->isVideoWebinar(),
             $happening->isWebinarRecorded(),
             $happening->isWebinarRecorded() && $happening->getEnd() < $this->dateTime,
-            $happening->getWebinarRecordZipFileUrl()
+            $happening->getWebinarRecordZipFileUrl(),
+            $happening->isPollAllowed()
         );
     }
 }