+++ CHANGELOG.md
@@ -5,9 +5,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2577 - Allow poll in webinar
+- VIMEET-2182 - UI for polls in webinar
+
+### Fixed
+- MV-318 - Conference order alphabetical
+
+## [2.95.0] - 2021-06-08
+### Added
+- VIMEET-2605 - Send follow up email only once with all met participants as receivers and user that made evaluation in carbon copy (cc)
+- VIMEET-2607 - Display all meeting participants in follow up email
+
+## [2.94.3] - 2021-06-08
+### Fixed
+- MV-339 - Display call visio button in the networking private chat
+
+## [2.94.2] - 2021-06-08
 ### Fixed
 - Hotfix - Improve exception info if meeting not found after evaluation timeout
-- MV-318 - Conference order alphabetical
+
+### Added
+- MV-341 - Fix exception if number of remaining priority meeting requests is negative
 
 ## [2.94.1] - 2021-06-03
 ### Fixed